requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic-settings==2.7.1
open-clip-torch==2.29.0
Pillow==11.1.0
python-dotenv==1.0.1
